refactor(invitation-system): extract error classes and validation to invitationValidation.ts
Moves all domain error classes (AccountAlreadyRegisteredError, ActivePendingError,
InvalidEmailError, InvalidRoleError, InvalidLinkError, ValidationError,
AccountCreatedConcurrentlyError, NotFoundError, NotPendingError, DecryptError)
and validation helpers (validateEmail, validateRole) plus the InvitationRole and
InvitationStatus type aliases from invitationStore.ts to the new
src/server/auth/invitationValidation.ts module.

invitationStore.ts now re-exports them for backward compatibility and imports them
locally. invitationController.ts imports error classes directly from
invitationValidation.ts.

Closes task lessons-from-luke-9c7.17.
